pyproject.toml
[tool.poetry]
name = "vorkers"
version = "0.1.0"
description = ""
authors = ["andrey-trasov <159887936+andrey-trasov@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
